Module holochain_common::paths

Constants

APPLICATION

DNA_EXTENSION

KEYS_DIRECTORY

N3H_BINARIES_DIRECTORY

ORGANIZATION

QUALIFIER

Functions

config_root
Returns the path to the root config directory for all of Holochain. If we can get a user directory it will be an XDG compliant path like "/home/peter/.config/holochain". If it can't get a user directory it will default to "/etc/holochain".

data_root
Returns the path to the root data directory for all of Holochain. If we can get a user directory it will be an XDG compliant path like "/home/peter/.local/share/holochain". If it can't get a user directory it will default to "/etc/holochain".

keys_directory
Returns the path to where agent keys are stored and looked for by default. Something like "~/.config/holochain/keys".

n3h_binaries_directory
Returns the path to where n3h binaries will be downloaded / run Something like "~/.local/share/holochain/n3h-binaries"

project_root
Returns the project root builder for holochain directories.